fixes #52
previously, preserveCamelCase was not treating integers as characters
with case. fix for issue #42 changed that, and characters that don't
have case are no longer considered to be capitalized. This exposed an
issue with integers acting as word breaks, which this commit should
resolve.

- no longer escapes early if lower characters followed by integers were
found (this was prematurely escaping sequences such as 'aa11aa'). The
escape early is no longer necessary, as the updated regex takes care of
it
- in the regex, integers are now considered affixed to the word preceding them, and
camel casing resumes after the completion of a series of integers. i.e.
'foo123bar' becomes 'foo123Bar' (before fixes for this issue and #42,
this wasn't occurring, but through the preserveCamelCase behavior and
the early escape mentioned above, we got this behavior)
- tests added!